requirements.txt
discord.py==2.5.2
pytz~=2024.1
flask~=3.0.2
aiopg~=1.4.0
python-dotenv~=1.0.0
requests~=2.32.3
asyncpg~=0.30.0
apscheduler~=3.10.4
sentry-sdk~=2.23.0
psycopg2-binary~=2.9.9
google-generativeai~=0.8.5
google~=3.0.0
protobuf~=5.29.5
pillow~=11.2.1
pynacl~=1.5.0
google-genai~=1.19.0
beautifulsoup4~=4.12.3